requirements.txt
aiohttp>=3.8.0
aiofiles>=0.8.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
pyyaml>=6.0
tldextract>=3.4.0